refactor raf
