+++ rpl_wei/processing/worker.py
@@ -72,23 +72,6 @@ def run_workflow_task(
     return result_payload
 
 
-def start_experiment(
-    experiment_name, experiment_id: Optional[Union[ulid.ULID, str]] = None
-):
-    log_dir = DATA_DIR / "runs" / experiment_id
-    result_dir = log_dir / "results"
-    exp_log = WEI_Logger.get_logger("log_" + str(experiment_id), log_dir)
-    exp_log.info(
-        "EXPERIMENT:START: "
-        + str(experiment_name)
-        + ", EXPERIMENT ID: "
-        + str(experiment_id)
-    )
-    log_dir.mkdir(parents=True, exist_ok=True)
-    result_dir.mkdir(parent=True, exist_ok=True)
-    return {"exp_dir": log_dir}
-
-
 if __name__ == "__main__":
     from rq import Worker
