+++ scripts/run_wei_server.sh
